requirements.txt
cnocr==2.2.3.2
Pillow==10.0.1
pynput==1.7.6
